requirements.txt
Pillow==10.4.0
streamlit==1.36.0
pandas==1.2.3
numpy==1.18.5
matplotlib==3.3.2
seaborn==0.11.0
scikit_learn==0.24.1
